backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3

# Time-series forecasting
prophet==1.1.5

# Machine learning
scikit-learn==1.3.2

# File uploads
python-multipart==0.0.6

# Environment variables
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# AI/ML
openai==1.3.0

# Testing
pytest==7.4.3
httpx==0.25.2